+++ src/macros.rs
@@ -1,7 +1,32 @@
 // macros.rs
 // This file is responsible for all the macros used by the kernel.
 
-// print!
+
+// block
+#[macro_export]
+macro_rules! block
+{
+	($e:expr) =>
+	{
+		loop
+		{
+			#[allow(unreachable_patterns)]
+			match $e
+			{
+				Err($crate::noblio::Err::Other(e)) =>
+				{
+					#[allow(unreachable_code)]
+					break Err(e)
+				}
+				Err($crate::noblkio::Err::WouldBlk) => {}
+				Ok(x) => break Ok(x),
+			}
+		}
+	};
+}
+
+
+// print
 //
 // This macro is analagous to the "print!" macro (std), except for the fact that this macro prints to the VGA text buffer.
 #[macro_export]
@@ -11,7 +36,7 @@ macro_rules! print
 }
 
 
-// println!
+// println
 //
 // This macro is analagous to the "println!" macro (std), except for the fact that this macro prints to the VGA text buffer.
 #[macro_export]