+++ src/math/int/mod.rs
@@ -1,0 +1,5 @@
+// src/math/intsr/mod.rs
+//
+// This is the mod.rs file for the math::int module, which handles the portion of the math module that pertains to integers.
+
+pub mod sr;