+++ src/math/mod.rs
@@ -1,0 +1,5 @@
+// math/mod.rs
+//
+// This is the mod.rs for the math module.
+
+pub mod int;